+++ Assets/Scripts/Presentation/Infrastructure/AttackBehaviour.cs
@@ -6,13 +6,13 @@ namespace Presentation.Infrastructure
 {
     public abstract class AttackBehaviour : ScriptableObject
     {
-        protected float MoneyToReceiveAfterHitEnemy;
+        protected int MoneyToReceiveAfterHitEnemy;
 
         public abstract void Init(MilitaryBuildingData militaryBuildingData);
 
         public abstract void DoAttack(Enemy receiveDamage);
 
-        public float GetMoneyOfAttack()
+        public int GetMoneyOfAttack()
         {
             return MoneyToReceiveAfterHitEnemy;
         }
